tested and fixed refactored rmdir options

+++ demo/src/main/java/os/cli/Main.java
@@ -331,13 +331,6 @@ public void touch(String com) { // 20220027
     // --------------------------- # Mahmoud Khaled 20220317 # --------------------------- //
 
     public void cd(String com) {
-        // System.out.println("cd called");
-        // System.out.println("args in comm: " + com);
-
-        // String[] MyArgs = proccess_args(com);
-        // for(int i = 1; i < MyArgs.length; i++) {
-        //     System.out.println(MyArgs[i]);
-        // }
 
         if ("..".equals(com)) {
             File newdir = new File(this.currentDir).getParentFile();
@@ -357,8 +350,6 @@ public void touch(String com) { // 20220027
 
             } else {
                 System.out.println("Directory " + com + " does not exists in " + this.currentDir);
-                // System.out.println(newdir.exists());
-                // System.out.println(newdir.isDirectory());
             }
 
         }
@@ -367,17 +358,12 @@ public void touch(String com) { // 20220027
 
     public void rmdir(String com) {                          //20220317
 
-        // System.out.println("rmdir called");
-        // System.out.println("args in comm: " + com);
-        // String[] MyArgs = proccess_args(com);
-        // for(int i = 1; i < MyArgs.length; i++) {
-        //     System.out.println(MyArgs[i]);
-        // }
         String[] Folders = proccess_args(com);
 
         HashMap<String, Integer> options = new HashMap<>();
 
         options.put("--ignore-fail-on-non-empty", 0);
+        options.put("-p", 0);
         options.put("-v", 0);
         options.put("--help", 0);
         options.put("--version", 0);
@@ -386,8 +372,49 @@ public void touch(String com) { // 20220027
             options.put(arg, 1);
         }
 
+        if(options.get("--help") == 1) {
+            System.out.println("""
+                Usage: rmdir [OPTION]... DIRECTORY...\r
+                Remove the DIRECTORY(ies), if they are empty.\r
+                \r
+                        --ignore-fail-on-non-empty\r
+                                ignore each failure that is solely because a directory\r
+                                is non-empty\r
+                    -p, --parents  remove DIRECTORY and its ancestors; e.g., 'rmdir -p a/b/c' is\r
+                                similar to 'rmdir a/b/c a/b a'\r
+                    -v, --verbose  output a diagnostic for every directory processed\r
+                        --help     display this help and exit\r
+                        --version  output version information and exit\r
+                \r
+                GNU coreutils online help: <https://www.gnu.org/software/coreutils/>\r
+                Report rmdir translation bugs to <https://translationproject.org/team/>\r
+                Full documentation at: <https://www.gnu.org/software/coreutils/rmdir>\r
+                or available locally via: info '(coreutils) rmdir invocation'\r
+                """ //
+            );
+            return;
+        }
+
+        if(options.get("--version") == 1) {
+            System.out.println("""
+                rmdir (GNU coreutils) X.Y\r
+                Copyright (C) YEAR Free Software Foundation, Inc.\r
+                License GPLv3+: GNU GPL version 3 or later <https://gnu.org/licenses/gpl.html>.\r
+                This is free software: you are free to change and redistribute it.\r
+                There is NO WARRANTY, to the extent permitted by law.\r
+                \r
+                Written by Mahmoud Khaled.\r
+                """ //
+            );
+            return;
+        }
+
 
         for (String Folder : Folders) {
+            if(Folder.charAt(0) == '-') {
+                continue;
+            }
+
             File folderToDelete = new File(this.currentDir, Folder);
 
             if(Folder.charAt(1) == ':') {
@@ -395,6 +422,7 @@ public void touch(String com) { // 20220027
             }
 
             if (!folderToDelete.exists()) {
+                System.out.println(folderToDelete.getPath());
                 System.out.println("Error: Folder does not exists.");
             } else if (!folderToDelete.isDirectory()) {
                 System.out.println("Error: Please provide a folder.");
@@ -403,56 +431,23 @@ public void touch(String com) { // 20220027
                     System.out.println("Error: Please provide an empty folder.");
                 }
             } else {
-                File FolderParent = new File(folderToDelete.getParent());
                 if (folderToDelete.delete()) {
+                    File FolderParent = new File(folderToDelete.getParent());
                     if(options.get("-v") == 1) {
                         System.out.println("Deleted Folder '" + folderToDelete.getName() + "' successfully");
                     }
-                    if (options.get("-p") == 1) {
-                       FolderParent.delete(); 
-                       if(options.get("-v") == 1) {
-                        System.out.println("Deleted Folder '" + FolderParent.getName() + "' successfully");
-                    }
+                    if (options.get("-p") == 1 && FolderParent.listFiles().length == 0) {
+                        FolderParent.delete(); 
+                        if(options.get("-v") == 1) {
+                            System.out.println("Deleted Folder '" + FolderParent.getName() + "' successfully");
+                        }   
                     }
                 } else {
                     System.out.println("Error: Failed to delete folder");
                 }
             }
 
-            if(options.get("--help") == 1) {
-                System.out.println("""
-                    Usage: rmdir [OPTION]... DIRECTORY...\r
-                    Remove the DIRECTORY(ies), if they are empty.\r
-                    \r
-                            --ignore-fail-on-non-empty\r
-                                    ignore each failure that is solely because a directory\r
-                                    is non-empty\r
-                        -p, --parents  remove DIRECTORY and its ancestors; e.g., 'rmdir -p a/b/c' is\r
-                                    similar to 'rmdir a/b/c a/b a'\r
-                        -v, --verbose  output a diagnostic for every directory processed\r
-                            --help     display this help and exit\r
-                            --version  output version information and exit\r
-                    \r
-                    GNU coreutils online help: <https://www.gnu.org/software/coreutils/>\r
-                    Report rmdir translation bugs to <https://translationproject.org/team/>\r
-                    Full documentation at: <https://www.gnu.org/software/coreutils/rmdir>\r
-                    or available locally via: info '(coreutils) rmdir invocation'\r
-                    """ //
-                );
-            }
-
-            if(options.get("--version") == 1) {
-                System.out.println("""
-                    rmdir (GNU coreutils) X.Y\r
-                    Copyright (C) YEAR Free Software Foundation, Inc.\r
-                    License GPLv3+: GNU GPL version 3 or later <https://gnu.org/licenses/gpl.html>.\r
-                    This is free software: you are free to change and redistribute it.\r
-                    There is NO WARRANTY, to the extent permitted by law.\r
-                    \r
-                    Written by Mahmoud Khaled.\r
-                    """ //
-                );
-            }
+            
         }
 
     }
@@ -468,6 +463,17 @@ public void touch(String com) { // 20220027
 
 
         String[] MyArgs = proccess_args(com);
+
+        HashMap<String, Integer> options = new HashMap<>();
+
+        options.put("--ignore-fail-on-non-empty", 0);
+        options.put("-v", 0);
+        options.put("--help", 0);
+        options.put("--version", 0);
+
+        for (String arg : MyArgs) {
+            options.put(arg, 1);
+        }
 
         for (String file : MyArgs) {
